+++ AlchemistCode/Cards/Uncommon/Resolve.cs
@@ -1,7 +1,6 @@
 using MegaCrit.Sts2.Core.Commands;
 using MegaCrit.Sts2.Core.Entities.Cards;
 using MegaCrit.Sts2.Core.GameActions.Multiplayer;
-using MegaCrit.Sts2.Core.Localization;
 using MegaCrit.Sts2.Core.Models.Powers;
 
 namespace Alchemist.AlchemistCode.Cards.Uncommon;
@@ -13,7 +12,7 @@ public class Resolve : AlchemistCard
 
     public Resolve() : base(1, CardType.Power, CardRarity.Uncommon, TargetType.Self)
     {
-        WithVar("Amount", 1, 0);
+        WithVar(new FermentVar("Amount", 1, "perTurn"));
         WithVar("perTurn", 1, 1);
         WithKeyword(CardKeyword.Retain);
         WithTip(typeof(StrengthPower));
@@ -22,11 +21,6 @@ public class Resolve : AlchemistCard
     private int Resolved => DynamicVars["Amount"].IntValue
         + DynamicVars["perTurn"].IntValue * FermentTurns;
 
-    protected override void AddExtraArgsToDescription(LocString description)
-    {
-        base.AddExtraArgsToDescription(description);
-        description.Add("Resolved", FermentTurns > 0 ? $" ([green]{Resolved}[/green])" : "");
-    }
 
     protected override async Task OnPlay(PlayerChoiceContext choiceContext, CardPlay play)
     {